Change in MainFrame to access relative directory path
Starting the program always gives a NullPointerException because the directory can't be found. This is because the default directory is given with an absolut path --> fileSelector.setCurrentFolder("f:/Code/FOnline/Fallout_Dat/art/critters");

Just changed this to a relative path to the directory, where the file(s) are. No other changes were done.